PGN Support added
Support for manipulation of files in PGN format was added. Also, was
added a class to represent games obtained from a PGN file, a class to
represent the moves of a game (a new instance of the Move interface with
new methods), and a utility class for manipulations with standard
algebraic notation and the notation used in the UCI protocol.